cdk/requirements.txt
aws-cdk-lib==2.201.0
constructs>=10.0.0,<11.0.0
python-dotenv>=0.20.0,<1.0.0
ruff>=0.12.1,<1.0.0